Cleaning

+++ mm/pmem.c
@@ -63,11 +63,11 @@ void pages_set(uint addr, uint size, bool reserved) {
 }
 
 void pmem_setup(void) {
-    mem_map = (struct page *)aligned(&_end);
+    mem_map = (struct page *)aligned((uint)&_end);
 
     struct memory_map *mapping = (struct memory_map *)mboot_mmap_addr();
 
-    int i;
+    uint i;
     for (i = 0; i < mboot_mmap_length(); ++i) {
         if (mapping[i].length_low == 0) continue;
